+++ R/data.R
@@ -20,6 +20,9 @@
 #' }
 "example_genes"
 
+#' @rdname example_genes
+"example_genes_polar"
+
 #' A set of example dummy alignment genes.
 #'
 #' Dummy genes, similar to those generated with `make_alignment_dummies()`, for
@@ -37,6 +40,9 @@
 #' @rdname example_genes
 "example_subgenes"
 
+#' @rdname example_genes
+"example_subgenes_polar"
+
 #' A set of example genetic features.
 #'
 #' Genetic features for example purposes only.
@@ -50,3 +56,21 @@
 #'   \item{forward}{is the feature oriented, and if so in the forward direction?}
 #' }
 "example_features"
+
+#' @rdname example_features
+"example_features_polar"
+
+#' A set of example terminators.
+#'
+#' Terminators for example purposes only.
+#'
+#' @format A data frame with two rows and three variables:
+#' \describe{
+#'   \item{molecule}{the genome}
+#'   \item{name}{the name of the terminator}
+#'   \item{position}{the position of the terminator}
+#' }
+"example_terminators"
+
+#' @rdname example_terminators
+"example_terminators_polar"