+++ R/geom_feature.R
@@ -23,7 +23,8 @@
 #' - alpha
 #' - colour
 #' - linetype
-#' - size
+#' - linewidth (the former size aesthetic has been deprecated and will be
+#' removed in future versions)
 #'
 #' @param mapping,data,stat,position,na.rm,show.legend,inherit.aes,... As
 #' standard for ggplot2. inherit.aes is set to FALSE by default, as features
@@ -47,7 +48,7 @@
 #'                                                      forward = forward)) +
 #'   ggplot2::facet_wrap(~ molecule, scales = "free")
 #'
-#' @seealso [geom_feature_label()]
+#' @seealso [geom_feature_label()], [geom_terminator()]
 #'
 #' @export
 geom_feature <- function(
@@ -90,7 +91,7 @@ GeomFeature <- ggplot2::ggproto("GeomFeature", ggplot2::Geom,
     alpha = 1,
     colour = "black",
     linetype = 1,
-    size = 1
+    linewidth = 1
   ),
 
   draw_key = ggplot2::draw_key_abline,
@@ -115,21 +116,23 @@ GeomFeature <- ggplot2::ggproto("GeomFeature", ggplot2::Geom,
     arrowhead_width
   ) {
 
-    # Detect flipped coordinates
-    coord_flip <- inherits(coord, "CoordFlip")
-
-    data <- coord$transform(data, panel_scales)
+    # Detect coordinate system and transform values
+    coord_system <- get_coord_system(coord)
+    data <- data_to_grid(data, coord_system, panel_scales, coord)
 
     gt <- grid::gTree(
       data = data,
-      cl = ifelse(coord_flip, "flipfeaturetree", "featuretree"),
+      cl = "featuretree",
       feature_height = feature_height,
       feature_width = feature_width,
-      arrowhead_width = arrowhead_width
+      arrowhead_width = arrowhead_width,
+      coord_system = coord_system
     )
     gt$name <- grid::grobName(gt, "geom_feature")
     gt
-  }
+  },
+  non_missing_aes = "size",
+  rename_size = TRUE
 )
 
 #' @importFrom grid makeContent
@@ -137,106 +140,59 @@ GeomFeature <- ggplot2::ggproto("GeomFeature", ggplot2::Geom,
 makeContent.featuretree <- function(x) {
 
   data <- x$data
-  feature_height <- x$feature_height
-  feature_width <- x$feature_width
-  arrowhead_width <- x$arrowhead_width
 
   # Prepare grob for each feature
   grobs <- lapply(seq_len(nrow(data)), function(i) {
 
     feature <- data[i, ]
 
+    # Set up geometry
+    r <- ifelse(x$coord_system == "polar", feature$away, NA)
+    feature_alongness <- unit_to_alaw(x$feature_width, "along", x$coord_system, r)
+    arrowhead_alongness <- unit_to_alaw(x$arrowhead_width, "along", x$coord_system, r)
+    feature_awayness <- unit_to_alaw(x$feature_height, "away", x$coord_system, r)
+
     # Determine whether this is a feature with orientation or not (i.e. whether
-    # or not to draw an elbow and arrowhead), and generate appropriate values
-    # for x-coordinates, y-coordinates and arrowhead
+    # or not to draw an elbow and arrowhead), and generate appropriate polyline
 
     # For non-oriented features:
     if (is.na(feature$forward) | ! is.logical(feature$forward)) {
 
-      end_y <- feature$y + grid::convertHeight(feature_height, "native", TRUE)
-      xs <- c(feature$x, feature$x)
-      ys <- c(feature$y, end_y)
+      alongs <- c(feature$along, feature$along)
+      aways <- c(feature$away, feature$away + feature_awayness)
       arrow <- NULL
 
     # For oriented features:
     } else {
 
       arrow_sign <- ifelse(feature$forward, 1, -1)
-      elbow_y <- feature$y + grid::convertHeight(feature_height, "native", TRUE)
-      end_x <- feature$x + (grid::convertWidth(feature_width, "native", TRUE) * arrow_sign)
-      xs <- c(feature$x, feature$x, end_x)
-      ys <- c(feature$y, elbow_y, elbow_y)
-      arrow <- grid::arrow(angle = 20, length = arrowhead_width, type = "closed")
+      end_along <- feature$along + (feature_alongness * arrow_sign)
+      alongs <- c(feature$along, feature$along, end_along)
+      aways <- c(feature$away, feature$away + feature_awayness, 
+                 feature$away + feature_awayness)
+      arrow <- grid::arrow(angle = 20, length = x$arrowhead_width, type = "closed")
     }
+
+    # If in polar coordinates, segment the polyline
+    if (x$coord_system == "polar") {
+      segmented <- segment_polarline(alongs, aways)
+      alongs <- segmented$thetas
+      aways <- segmented$rs
+    }
+
+    # Convert polyline into Cartesian coordinates within the grid viewport
+    coords <- alaw_to_grid(alongs, aways, x$coord_system, r)
     
     # Generate polyline grob for the feature
     pg <- grid::polylineGrob(
-      x = xs,
-      y = ys,
+      x = coords$x,
+      y = coords$y,
       arrow = arrow,
       gp = grid::gpar(
         col = feature$colour,
         fill = feature$colour,
         lty = feature$linetype,
-        lwd = feature$size
-      )
-    )
-
-    # Return the grob
-    pg
-  })
-
-  class(grobs) <- "gList"
-  grid::setChildren(x, grobs)
-}
-
-#' @importFrom grid makeContent
-#' @export
-makeContent.flipfeaturetree <- function(x) {
-
-  data <- x$data
-  feature_height <- x$feature_height
-  feature_width <- x$feature_width
-  arrowhead_width <- x$arrowhead_width
-
-  # Prepare grob for each feature
-  grobs <- lapply(seq_len(nrow(data)), function(i) {
-
-    feature <- data[i, ]
-
-    # Determine whether this is a feature with orientation or not (i.e. whether
-    # or not to draw an elbow and arrowhead), and generate appropriate values
-    # for x-coordinates, y-coordinates and arrowhead
-
-    # For non-oriented features:
-    if (is.na(feature$forward) | ! is.logical(feature$forward)) {
-
-      end_x <- feature$x + grid::convertWidth(feature_height, "native", TRUE)
-      ys <- c(feature$y, feature$y)
-      xs <- c(feature$x, end_x)
-      arrow <- NULL
-
-    # For oriented features:
-    } else {
-
-      arrow_sign <- ifelse(feature$forward, 1, -1)
-      elbow_x <- feature$x + grid::convertWidth(feature_height, "native", TRUE)
-      end_y <- feature$y + (grid::convertHeight(feature_width, "native", TRUE) * arrow_sign)
-      ys <- c(feature$y, feature$y, end_y)
-      xs <- c(feature$x, elbow_x, elbow_x)
-      arrow <- grid::arrow(angle = 20, length = arrowhead_width, type = "closed")
-    }
-    
-    # Generate polyline grob for the feature
-    pg <- grid::polylineGrob(
-      y = ys,
-      x = xs,
-      arrow = arrow,
-      gp = grid::gpar(
-        col = feature$colour,
-        fill = feature$colour,
-        lty = feature$linetype,
-        lwd = feature$size
+        lwd = (feature$linewidth %||% feature$size)
       )
     )